管理员登入自动进入后台管理页面；完善发帖管理和举报管理

+++ uclub-backend/src/main/java/com/example/uclub_backend/forum/controller/PostController.java
@@ -184,7 +184,7 @@ public ResponseEntity<Map<String, Object>> toggleLike(@PathVariable Long id, @Re
     
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deletePost(@PathVariable Long id, @RequestParam Long userId) {
+    public ResponseEntity<?> deletePost(@PathVariable Long id, @RequestParam Integer userId) {
         try {
             postService.deletePostById(id, userId);
             return ResponseEntity.ok(Map.of("code", 200, "message", "删除成功"));
